ref(useWeb3Modal.ts): simplify modal and move logic to useWagmi
